Sean Alpha Feedback
Overall, I think it's a good start but there's a lot of work to be done. I think some core mechanics are missing, and you need to get the PlayerController working properly in the upcoming week even before the Beta. I would focus on that programming wise, and then make sure you have decided on a couple of other enemy and platform types.
Alpha grade: 80/100